backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.4
anthropic==0.43.0
# ⚠️ 임시: 개발/테스트 단계에서 OpenAI API 사용. 최종 통합 시 제거 가능.
openai==1.57.0
httpx==0.28.1
PyMuPDF==1.25.1
openpyxl==3.1.5
# 라벨 이미지 크롭 & DOCX 파일 OCR
Pillow==12.1.0
python-docx==1.1.2
# OCR 분석 결과 문서 내보내기 (DOCX/PDF)
reportlab==4.2.5
python-dotenv==1.0.1
supabase==2.11.0
# F1: asyncpg 커넥션 풀 (직접 SQL 쿼리)
asyncpg>=0.29.0
# F2/F3/F4/F5: Pinecone 벡터 검색
pinecone-client>=5.0.0
# F4: multilingual-e5-large 임베딩
sentence-transformers>=2.2.0
# F4: PDF 레포트 생성
fpdf2>=2.7.0
# F5: Voyage-3 임베딩
voyageai>=0.3.0
